requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aethersprite @ git+https://github.com/haliphax/aethersprite.git@main
    # via -r requirements.in
aiohappyeyeballs==2.3.5
    # via
    #   aethersprite
    #   aiohttp
aiohttp==3.10.11
    # via
    #   aethersprite
    #   discord-py
aiosignal==1.3.1
    # via
    #   aethersprite
    #   aiohttp
annotated-types==0.6.0
    # via
    #   aethersprite
    #   pydantic
anyio==4.3.0
    # via
    #   aethersprite
    #   starlette
    #   watchfiles
attrs==23.2.0
    # via
    #   aethersprite
    #   aiohttp
click==8.1.7
    # via
    #   aethersprite
    #   uvicorn
colorlog==6.8.2
    # via aethersprite
discord-pretty-help==2.0.7
    # via aethersprite
discord-py==2.3.2
    # via aethersprite
fastapi==0.110.0
    # via aethersprite
frozenlist==1.4.1
    # via
    #   aethersprite
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via
    #   aethersprite
    #   uvicorn
hiredis==3.1.0
    # via -r requirements.in
httptools==0.6.1
    # via
    #   aethersprite
    #   uvicorn
idna==3.7
    # via
    #   aethersprite
    #   anyio
    #   yarl
multidict==6.0.5
    # via
    #   aethersprite
    #   aiohttp
    #   yarl
propcache==0.2.0
    # via
    #   aethersprite
    #   yarl
pydantic==2.6.4
    # via
    #   aethersprite
    #   fastapi
    #   realm-schema
pydantic-core==2.16.3
    # via
    #   aethersprite
    #   pydantic
python-dotenv==1.0.1
    # via
    #   aethersprite
    #   uvicorn
pytz==2024.1
    # via aethersprite
pyyaml==6.0.1
    # via
    #   aethersprite
    #   uvicorn
realm-schema @ git+https://github.com/realm-ttrpg/schema.git@main
    # via -r requirements.in
redis==5.2.1
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   aethersprite
    #   anyio
sqlitedict==2.1.0
    # via aethersprite
starlette==0.36.3
    # via
    #   aethersprite
    #   fastapi
toml==0.10.2
    # via aethersprite
typing-extensions==4.10.0
    # via
    #   aethersprite
    #   fastapi
    #   pydantic
    #   pydantic-core
uvicorn[standard]==0.29.0
    # via aethersprite
uvloop==0.19.0
    # via
    #   aethersprite
    #   uvicorn
watchfiles==0.21.0
    # via
    #   aethersprite
    #   uvicorn
websockets==12.0
    # via
    #   aethersprite
    #   uvicorn
yarl==1.17.2
    # via
    #   aethersprite
    #   aiohttp
